refactor: model package명을 domain package로 이름을 변경
