+++ src/main/java/baseball/domain/BaseballGameCalculator.java
@@ -1,4 +1,4 @@
-package baseball.model;
+package baseball.domain;
 
 import java.util.List;
