+++ src/main/java/baseball/domain/ComBaseballNumber.java
@@ -1,4 +1,4 @@
-package baseball.model;
+package baseball.domain;
 
 import java.util.ArrayList;
 import java.util.List;
