+++ src/main/java/baseball/domain/UserBaseballNumber.java
@@ -1,4 +1,4 @@
-package baseball.model;
+package baseball.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
